@@ -23,7 +23,6 @@ SOFTWARE.
 package config
 
 import (
-	"io/ioutil"
 	"os"
 	"testing"
 
@@ -57,7 +56,7 @@ func WriteFile(t *testing.T, s string) string {
 	current, err := os.Getwd()
 	require.NoError(t, err)
 
-	file, err := ioutil.TempFile(current, "*")
+	file, err := os.CreateTemp(current, "*")
 	require.NoError(t, err)
 
 	_, err = file.WriteString(s)
